requirements.txt
allure-pytest==2.13.2
allure-python-commons==2.13.2
astroid==2.11.7
async-generator==1.10
attrs==23.1.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
dill==0.3.6
exceptiongroup==1.1.1
flake8==5.0.4
h11==0.14.0
idna==3.4
importlib-metadata==4.2.0
iniconfig==2.0.0
isort==5.11.5
lazy-object-proxy==1.9.0
mccabe==0.7.0
outcome==1.2.0
packaging==23.1
pep8==1.7.1
platformdirs==3.5.1
pluggy==1.0.0
pycodestyle==2.9.1
pycparser==2.21
pyflakes==2.5.0
pylint==2.13.9
PySocks==1.7.1
pytest==7.3.1
pytest-ordering==0.6
python-dotenv==0.21.1
requests==2.30.0
selenium==4.9.0
sniffio==1.3.0
sortedcontainers==2.4.0
tomli==2.0.1
tqdm==4.65.0
trio==0.22.0
trio-websocket==0.10.2
typed-ast==1.5.4
typing_extensions==4.5.0
urllib3==1.26.15
webdriver-manager==3.8.6
wrapt==1.15.0
wsproto==1.2.0
zipp==3.15.0
